+++ structurizr-autolayout/src/test/java/com/structurizr/autolayout/graphviz/DOTExporterTests.java
@@ -79,7 +79,7 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeSystemLandscapeViewWithGroupedElements() throws Exception {
+    public void test_writeSystemLandscapeViewWithGroupedElements() {
         Workspace workspace = new Workspace("Name", "");
         CustomElement box = workspace.getModel().addCustomElement("Box");
         Person user = workspace.getModel().addPerson("User", "");
@@ -121,7 +121,50 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeSystemLandscapeViewWithNestedGroupedElements() throws Exception {
+    public void test_writeSystemLandscapeViewWithGroupedElementsAndGroupPadding() {
+        Workspace workspace = new Workspace("Name", "");
+        CustomElement box = workspace.getModel().addCustomElement("Box");
+        Person user = workspace.getModel().addPerson("User", "");
+        user.setGroup("External");
+        SoftwareSystem softwareSystem = workspace.getModel().addSoftwareSystem("Software System", "");
+        softwareSystem.setGroup("Internal");
+        user.uses(softwareSystem, "Uses");
+
+        SystemLandscapeView view = workspace.getViews().createSystemLandscapeView("SystemLandscape", "");
+        view.addAllElements();
+        view.add(box);
+        view.addProperty("structurizr.groupPadding", "50");
+
+        DOTExporter exporter = new DOTExporter(RankDirection.TopBottom, 300, 300);
+        Diagram diagram = exporter.export(view);
+
+        String content = diagram.getDefinition();
+        assertEquals("""
+                digraph {
+                  compound=true
+                  graph [splines=polyline,rankdir=TB,ranksep=1.0,nodesep=1.0,fontsize=5]
+                  node [shape=box,fontsize=5]
+                  edge []
+                
+                  subgraph "cluster_group_1" {
+                    margin=50
+                    2 [width=1.500000,height=1.000000,fixedsize=true,id=2,label="2: User"]
+                  }
+                
+                  subgraph "cluster_group_2" {
+                    margin=50
+                    3 [width=1.500000,height=1.000000,fixedsize=true,id=3,label="3: Software System"]
+                  }
+                
+                  1 [width=1.500000,height=1.000000,fixedsize=true,id=1,label="1: Box"]
+                
+                  2 -> 3 [id=4]
+                
+                }""", content);
+    }
+
+    @Test
+    public void test_writeSystemLandscapeViewWithNestedGroupedElements() {
         Workspace workspace = new Workspace("Name", "");
         workspace.getModel().addProperty("structurizr.groupSeparator", "/");
 
@@ -183,7 +226,7 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeSystemLandscapeViewInGermanLocale() throws Exception {
+    public void test_writeSystemLandscapeViewInGermanLocale() {
         // ranksep=1.0 was being output as ranksep=1,0
         Locale.setDefault(new Locale("de", "DE"));
         Workspace workspace = new Workspace("Name", "");
@@ -217,7 +260,7 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeSystemContextView() throws Exception {
+    public void test_writeSystemContextView() {
         Workspace workspace = new Workspace("Name", "");
         CustomElement box = workspace.getModel().addCustomElement("Box");
         Person user = workspace.getModel().addPerson("User", "");
@@ -250,7 +293,7 @@ public class DOTExporterTests {
 
 
     @Test
-    public void test_writeSystemContextViewWithGroupedElements() throws Exception {
+    public void test_writeSystemContextViewWithGroupedElements() {
         Workspace workspace = new Workspace("Name", "");
         CustomElement box = workspace.getModel().addCustomElement("Box");
         Person user = workspace.getModel().addPerson("User", "");
@@ -292,7 +335,7 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeContainerViewWithGroupedElementsInASingleSoftwareSystem() throws Exception {
+    public void test_writeContainerViewWithGroupedElementsInASingleSoftwareSystem() {
         Workspace workspace = new Workspace("Name", "");
         CustomElement box = workspace.getModel().addCustomElement("Box");
         SoftwareSystem softwareSystem = workspace.getModel().addSoftwareSystem("Software System", "");
@@ -344,7 +387,7 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeContainerViewWithGroupedElementsInMultipleSoftwareSystems() throws Exception {
+    public void test_writeContainerViewWithGroupedElementsInMultipleSoftwareSystems() {
         Workspace workspace = new Workspace("Name", "");
 
         SoftwareSystem softwareSystem1 = workspace.getModel().addSoftwareSystem("Software System 1");
@@ -396,7 +439,36 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeComponentViewWithGroupedElements() throws Exception {
+    public void test_writeContainerViewWithBoundaryPadding() {
+        Workspace workspace = new Workspace("Name", "");
+        SoftwareSystem softwareSystem = workspace.getModel().addSoftwareSystem("Software System", "");
+        softwareSystem.addContainer("Container");
+
+        ContainerView view = workspace.getViews().createContainerView(softwareSystem, "key");
+        view.addAllElements();
+        workspace.getViews().getConfiguration().addProperty("structurizr.boundaryPadding", "50");
+
+        DOTExporter exporter = new DOTExporter(RankDirection.TopBottom, 300, 300);
+        Diagram diagram = exporter.export(view);
+
+        String content = diagram.getDefinition();
+        assertEquals("""
+                digraph {
+                  compound=true
+                  graph [splines=polyline,rankdir=TB,ranksep=1.0,nodesep=1.0,fontsize=5]
+                  node [shape=box,fontsize=5]
+                  edge []
+                
+                  subgraph cluster_1 {
+                    margin=50
+                    2 [width=1.500000,height=1.000000,fixedsize=true,id=2,label="2: Container"]
+                  }
+                
+                }""", content);
+    }
+
+    @Test
+    public void test_writeComponentViewWithGroupedElements() {
         Workspace workspace = new Workspace("Name", "");
         CustomElement box = workspace.getModel().addCustomElement("Box");
         SoftwareSystem softwareSystem = workspace.getModel().addSoftwareSystem("Software System", "");
@@ -449,7 +521,7 @@ public class DOTExporterTests {
     }
 
     @Test
-    public void test_writeContainerViewWithGroupedElements_WithAndWithoutAGroupSeparator() throws Exception {
+    public void test_writeContainerViewWithGroupedElements_WithAndWithoutAGroupSeparator() {
         Workspace workspace = new Workspace("Name", "");
         SoftwareSystem softwareSystem = workspace.getModel().addSoftwareSystem("Software System", "");
         Container container1 = softwareSystem.addContainer("Container 1");
@@ -530,6 +602,55 @@ public class DOTExporterTests {
                         margin=25
                         subgraph cluster_10 {
                           margin=25
+                          11 [width=1.500000,height=1.000000,fixedsize=true,id=11,label="11: Web Application"]
+                        }
+                
+                      }
+                
+                    }
+                
+                  }
+                
+                  11 -> 14 [id=15]
+                  7 -> 8 [id=16]
+                  8 -> 11 [id=17]
+                
+                }""", diagram.getDefinition());
+    }
+
+    @Test
+    public void test_AmazonWebServicesExampleWithDeploymentNodePadding() throws Exception {
+        Workspace workspace = WorkspaceUtils.loadWorkspaceFromJson(new File("src/test/resources/structurizr-54915-workspace.json"));
+        workspace.getViews().getConfiguration().addProperty("structurizr.deploymentNodePadding", "50");
+        DOTExporter exporter = new DOTExporter(RankDirection.LeftRight, 300, 300);
+        Diagram diagram = exporter.export(workspace.getViews().getDeploymentViews().iterator().next());
+
+        assertEquals("""
+                digraph {
+                  compound=true
+                  graph [splines=polyline,rankdir=LR,ranksep=1.0,nodesep=1.0,fontsize=5]
+                  node [shape=box,fontsize=5]
+                  edge []
+                
+                  subgraph cluster_5 {
+                    margin=50
+                    subgraph cluster_6 {
+                      margin=50
+                      subgraph cluster_12 {
+                        margin=50
+                        subgraph cluster_13 {
+                          margin=50
+                          14 [width=1.500000,height=1.000000,fixedsize=true,id=14,label="14: Database"]
+                        }
+                
+                      }
+                
+                      7 [width=1.500000,height=1.000000,fixedsize=true,id=7,label="7: Route 53"]
+                      8 [width=1.500000,height=1.000000,fixedsize=true,id=8,label="8: Elastic Load Balancer"]
+                      subgraph cluster_9 {
+                        margin=50
+                        subgraph cluster_10 {
+                          margin=50
                           11 [width=1.500000,height=1.000000,fixedsize=true,id=11,label="11: Web Application"]
                         }